I have no idea

+++ app/src/main/java/is/fb01/tud/university/mobilesurveystud/BackEnd/Service/HandlerService/HandlerDetectorServiceBase.java
@@ -40,7 +40,7 @@ public abstract class HandlerDetectorServiceBase extends DetectorServiceBase {
         super.onDestroy();
     }
 
-    abstract int getHandlerDelay();
+    abstract long getHandlerDelay();
     abstract boolean conditionToCheck();
 
     private void runHandler() {
