+++ app/src/main/java/is/fb01/tud/university/mobilesurveystud/BackEnd/Service/HandlerService/SoundDetectionService.java
@@ -31,7 +31,7 @@ public class SoundDetectionService extends HandlerDetectorServiceBase{
     }
 
     @Override
-    int getHandlerDelay() {
+    long getHandlerDelay() {
         return GlobalSettings.gSoundRequestWait;
     }
